Fix Datepicker to persist selected relative range
