+++ src/actions/suppressions.js
@@ -1,4 +1,11 @@
+import moment from 'moment';
+import config from 'src/config';
+import _ from 'lodash';
 import sparkpostApiRequest from 'src/actions/helpers/sparkpostApiRequest';
+import { refreshReportRange } from 'src/actions/reportFilters';
+
+const { apiDateFormat } = config;
+
 
 export function checkSuppression() { //used in DashBoardPage to check if account has suppression
   const params = { sources: 'Manually Added', limit: 1 };
@@ -33,9 +40,30 @@ export function searchRecipient({ email, subaccountId } = {}) {
     );
 }
 
-export function listSuppressions(params) {
+export function listSuppressions(options) {
+  const { from, to, types = [], sources = []} = options;
+  const params = {};
 
-  return (dispatch, getState) => dispatch(
+  if (from) {
+    params.from = moment(options.from).utc().format(apiDateFormat);
+  }
+
+  if (to) {
+    params.to = moment(options.to).utc().format(apiDateFormat);
+  }
+
+  if (!_.isEmpty(types)) {
+    params.types = types.join(',');
+  }
+
+  if (!_.isEmpty(sources)) {
+    params.sources = sources.join(',');
+  }
+
+  return (dispatch, getState) => {
+    dispatch(refreshReportRange(options));
+
+    return dispatch(
       sparkpostApiRequest({
         type: 'GET_SUPPRESSIONS',
         meta: {
@@ -45,4 +73,5 @@ export function listSuppressions(params) {
         }
       })
     );
+  };
 }
